chore(deleting-files): refactor deleting files in any endpoint
