+++ src/services/storyService.ts
@@ -2,8 +2,7 @@ import AppError from '@base/errors/AppError';
 import Story from '@base/models/storyModel';
 import User from '@base/models/userModel';
 import mongoose from 'mongoose';
-import fs from 'fs/promises';
-import path from 'path';
+import deleteFile from '@base/utils/deleteFile';
 import Chat from '@base/models/chatModel';
 import IStory from '@base/types/story';
 
@@ -41,21 +40,7 @@ export const deleteStoryFile = async (storyId: mongoose.Types.ObjectId) => {
   }
 
   const fileName = story.content;
-  if (!fileName || !fileName.trim()) return;
-
-  const filePath = path.join(process.cwd(), 'src/public/media/', fileName);
-
-  try {
-    // Check if the file exists
-    await fs.access(filePath);
-    // Delete the file if it exists
-    await fs.unlink(filePath);
-  } catch (err: any) {
-    // Ignore file not found errors (ENOENT)
-    if (err.code !== 'ENOENT') {
-      throw err; // Rethrow unexpected errors
-    }
-  }
+  await deleteFile(fileName);
 };
 
 // Returns all the users ids that the user has a private chat with
